APIS-5829 - notification service refactor
