+++ app/uk/gov/hmrc/thirdpartyapplication/services/ApplicationUpdateService.scala
@@ -16,14 +16,14 @@
 
 package uk.gov.hmrc.thirdpartyapplication.services
 
-import cats.data.{EitherT, NonEmptyChain, NonEmptyList, Validated}
+import cats.data.{EitherT, NonEmptyChain, Validated}
 import uk.gov.hmrc.apiplatform.modules.common.services.{ApplicationLogger, EitherTHelper}
 import uk.gov.hmrc.thirdpartyapplication.domain.models._
 import uk.gov.hmrc.thirdpartyapplication.models.db.ApplicationData
 import uk.gov.hmrc.thirdpartyapplication.repository.ApplicationRepository
 import uk.gov.hmrc.thirdpartyapplication.services.commands._
 import uk.gov.hmrc.http.HeaderCarrier
-
+import uk.gov.hmrc.thirdpartyapplication.services.notifications.NotificationService
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.{ExecutionContext, Future}
